+++ include/flexflow/ops/softmax.h
@@ -1,32 +1,12 @@
 #ifndef _FLEXFLOW_SOFTMAX_H
 #define _FLEXFLOW_SOFTMAX_H
 
-#include "flexflow/device.h"
-#include "flexflow/fftype.h"
 #include "flexflow/layer.h"
 #include "flexflow/node.h"
-#include "flexflow/op_meta.h"
 #include "flexflow/operator.h"
 #include "flexflow/ops/softmax_params.h"
 
 namespace FlexFlow {
-
-class Softmax;
-
-class SoftmaxMeta : public OpMeta {
-public:
-  SoftmaxMeta(FFHandler handle,
-              Softmax const *softmax,
-              Legion::Domain const &input_domain);
-#if defined(FF_USE_CUDA) || defined(FF_USE_HIP_CUDA)
-  cudnnTensorDescriptor_t inputTensor;
-#else
-  miopenTensorDescriptor_t inputTensor;
-#endif
-  bool profiling;
-  int dim;
-  char op_name[MAX_OPNAME];
-};
 
 class Softmax : public Op {
 public:
@@ -63,28 +43,9 @@ public:
                             std::vector<Legion::PhysicalRegion> const &regions,
                             Legion::Context ctx,
                             Legion::Runtime *runtime);
-  void init_meta(SoftmaxMeta *m,
-                 Legion::Rect<2> const &input,
-                 Legion::Rect<2> const &output) const;
   bool measure_operator_cost(Simulator *sim,
                              MachineView const &pc,
                              CostMetrics &cost_metrics) const override;
-  static void forward_kernel(SoftmaxMeta const *m,
-                             float const *input_ptr,
-                             float *output_ptr,
-                             ffStream_t stream);
-  static void forward_kernel_wrapper(SoftmaxMeta const *m,
-                                     float const *input_ptr,
-                                     float *output_ptr);
-  static void backward_kernel(float *input_grad_ptr,
-                              float const *output_grad_ptr,
-                              size_t num_elements,
-                              ffStream_t stream);
-  static void backward_kernel_wrapper(SoftmaxMeta const *m,
-                                      float *input_grad_ptr,
-                                      float const *output_grad_ptr,
-                                      size_t num_elements);
-
   Params get_params() const;
 
 private:
