+++ src/ops/fused.cu
@@ -24,10 +24,10 @@
 #include "flexflow/ops/kernels/concat_kernels.h"
 #include "flexflow/ops/kernels/conv_2d_kernels.h"
 #include "flexflow/ops/kernels/element_binary_kernels.h"
+#include "flexflow/ops/kernels/flat_kernels.h"
 #include "flexflow/ops/kernels/linear_kernels.h"
 #include "flexflow/ops/kernels/pool_2d_kernels.h"
 #include "flexflow/ops/kernels/reshape_kernels.h"
-#include "flexflow/ops/kernels/flat_kernels.h"
 #include "flexflow/ops/kernels/transpose_kernels.h"
 #include "flexflow/utils/cuda_helper.h"
